consolidate defaultStates
