+++ src/game/store/defaults/defaultGameStoreState.ts
@@ -1,0 +1,24 @@
+import { GameStoreState, DoorDir } from '../types';
+import getDefaultGameState from './getDefaultGameState';
+
+const defaultGameStoreState: GameStoreState = {
+  transition: {
+    dest: null,
+    dir: DoorDir.FORWARD,
+  },
+  transient: {
+    nextText: '',
+    nextMusic: {
+      fileName: null,
+    },
+  },
+  text: null,
+  loading: true,
+  cursorEnabled: false,
+  gameName: '',
+  menu: 'MAIN',
+  images: new Map(),
+  ...getDefaultGameState(''),
+};
+
+export default defaultGameStoreState;
