+++ src/game/store/defaults/defaultVerbs.ts
@@ -1,0 +1,40 @@
+const defaultVerbs = [
+  {
+    name: 'MOVE',
+    defaultText: 'blah',
+  },
+  {
+    name: 'LOOK',
+    defaultText: 'blah',
+  },
+  {
+    name: 'OPEN',
+    defaultText: 'blah',
+  },
+  {
+    name: 'USE',
+    defaultText: 'blah',
+  },
+  {
+    name: 'SMOKE',
+    defaultText: 'Smoking that would be ill-advised!',
+  },
+  {
+    name: 'TAKE',
+    defaultText: 'blah',
+  },
+  {
+    name: 'HIT',
+    defaultText: 'Ya blew it. That really hurt.',
+  },
+  {
+    name: 'EAT',
+    defaultText: "Don't eat that.",
+  },
+  {
+    name: 'SPEAK',
+    defaultText: 'No response.',
+  },
+];
+
+export default defaultVerbs;
